# api/docs/swagger.yaml
definitions:
  models.Car:
    properties:
      brand:
        type: string
      colour:
        type: string
      createdAt:
        type: string
      engineCap:
        type: number
      hoursepower:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      order:
        items:
          $ref: '#/definitions/models.GetOrder'
        type: array
      updatedAt:
        type: string
      year:
        type: integer
    type: object
  models.CarReal:
    properties:
      brand:
        type: string
      colour:
        type: string
      createdAt:
        type: string
      engine_cap:
        type: number
      hourse_power:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      year:
        type: integer
    type: object
  models.CreateCar:
    properties:
      brand:
        type: string
      colour:
        type: string
      createdAt:
        type: string
      engineCap:
        type: number
      hoursepower:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      year:
        type: integer
    type: object
...
  models.Customer:
    properties:
      createdAt:
        type: string
      first_name:
        type: string
      gmail:
        type: string
      id:
        type: string
      isblocked:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
  models.CustomerLoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  models.CustomerLoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  models.CustomerRegisterRequest:
    properties:
      mail:
        type: string
    type: object
  models.GetAllCarsRealResponse:
    properties:
      cars:
        items:
          $ref: '#/definitions/models.CarReal'
        type: array
      count:
        type: integer
    type: object
  models.GetAllCarsResponse:
    properties:
      cars:
        items:
          $ref: '#/definitions/models.Car'
        type: array
      count:
        type: integer
    type: object
...
  models.GetByIDCar:
    properties:
      brand:
        type: string
      colour:
        type: string
      createdAt:
        type: string
      engineCap:
        type: number
      hoursepower:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      year:
        type: integer
    type: object
  models.GetOrder:
    properties:
      amount:
        type: number
      car:
        $ref: '#/definitions/models.Car'
      created_at:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      from_date:
        type: string
      id:
        type: string
      paid:
        type: boolean
      status:
        type: string
      to_date:
        type: string
      updated_at:
        type: string
    type: object
  models.LoginCustomer:
    properties:
      first_name:
        type: string
      gmail:
        type: string
      gmailcode:
        type: string
      last_name:
        type: string
      login:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.UpdateCar:
    properties:
      brand:
        type: string
      colour:
        type: string
      createdAt:
        type: string
      engineCap:
        type: number
      hoursepower:
        type: integer
      id:
        type: string
      model:
        type: string
      name:
        type: string
      updatedAt:
        type: string
      year:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /availablecars:
    get:
      consumes:
      - application/json
      description: get availablecars
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GetAllCarsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get availablecars
      tags:
      - car
  /car:
    post:
      consumes:
      - application/json
      description: create a new car
      parameters:
      - description: car
        in: body
        name: car
        schema:
          $ref: '#/definitions/models.CreateCar'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateCar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Creates a new cars
      tags:
      - car
  /car/{id}:
    delete:
      consumes:
      - application/json
      description: Delete car
      parameters:
      - description: car_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete car
      tags:
      - car
    get:
      consumes:
      - application/json
      description: get car by ID
      parameters:
      - description: car
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GetByIDCar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Gets a car by ID
      tags:
      - car
    put:
      consumes:
      - application/json
      description: Updating car
      parameters:
      - description: car_id
        in: path
        name: id
        required: true
        type: string
      - description: car
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCar'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UpdateCar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Updating car
      tags:
      - car
  /cars:
    get:
      consumes:
      - application/json
      description: get car list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GetAllCarsRealResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get car list
      tags:
      - car
...
  /customer/auth/create:
    post:
      consumes:
      - application/json
      description: Customer password check and create
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginCustomer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LoginCustomer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Customer password check and create
      tags:
      - auth
  /customer/login:
    post:
      consumes:
      - application/json
      description: Customer login
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.CustomerLoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CustomerLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Customer login
      tags:
      - auth
...
  /customer/register:
    post:
      consumes:
      - application/json
      description: Customer register
      parameters:
      - description: register
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models.CustomerRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Customer register
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"